Added the docs and passed the test.
